Lesson status page
Started the lesson status page. I have a visual of what I was aiming for.
Having trouble with displaying and aligning things. I ignored the functionalities at the moment as I wasn't sure what to connect it to but this is the basic layout I thought of based on the old site.