Deleted diviuvi.JPG and drenglish.png
As Min replicated the design of the profile cards into HTML and CSS then Greg and Simon generated the proper favicons and Toby made the navbar version of the logo, these images were deleted. Also Andrei, diviuvi means important and strong in a language you don't know so it's not a nonsensical name.